Added options
I needed to specify these options. All of the tests pass, but I have not written new ones. Let me know if you want me to before merge, happy to - but I do not want you to fee like you have to maintain this module if it was just an example for me. thank you again!